+++ megalinter/pre_post_factory.py
@@ -18,13 +18,32 @@ def run_post_commands(mega_linter):
     return run_pre_post_commands("POST_COMMANDS", "[Post]", mega_linter)
 
 
+# Commands to run before a linter (defined in descriptors)
+def run_linter_pre_commands(mega_linter, linter):
+    if linter.pre_commands is not None:
+        return run_commands(linter.pre_commands, "[Pre]["+linter.name+"]" , mega_linter)
+    return []
+
+
+# Commands to run before a linter (defined in descriptors)
+def run_linter_post_commands(mega_linter, linter):
+    if linter.post_commands is not None:
+        return run_commands(linter.post_commands, "[Post]["+linter.name+"]" , mega_linter)
+    return []
+
+
+# Get commands from configuration
 def run_pre_post_commands(key, log_key, mega_linter):
-    pre_commands = config.get_list(key, None)
+    pre_or_post_commands = config.get_list(key, None)
+    return run_commands(pre_or_post_commands, log_key, mega_linter)
+
+# Perform run of commands
+def run_commands(all_commands, log_key, mega_linter):
     pre_commands_results = []
-    if pre_commands is None:
+    if all_commands is None:
         logging.debug(f"{log_key} No commands declared in user configuration")
         return pre_commands_results
-    for command_info in pre_commands:
+    for command_info in all_commands:
         pre_command_result = run_command(command_info, log_key, mega_linter)
         pre_commands_results += [pre_command_result]
     return pre_commands_results